Updated and refactored message rendering (#78)
* Moved upload and link-preview rendering to template.

Removed upload and link-preview rendering from the Message class, and
replaced it with wrapper code to expose the interesting parts of those
objects to the template.

Modified template and CSS to allow presenting the current attachment and
upload format, with all the HTML generation now done in templates instead
of in python.

* Made thumbnail generation more robust.

Made thumbnail generation succeed if there is a special key based on
the file type (e.g. "thumb_pdf") instead of the normal set of numeric
keys, and also fall back to a random numeric thumbnail key if the
requested one is not present.

* Fixed regex for channel ID references.

Channel ID variable section no longer necessarily starts with a 0.

* Made the User object more bot-friendly.

Bot users (especially those that have to be created on the fly) may not
have all the fields that normal users do: made the User object tolerate
this behavior.

* Front-end changes for better attachments/bots.

Updated the message-rendering template and CSS to handle more complicated attachments,
such as are produced by certain popular integrations (looking at you, Jenkins).

* Reworked markup/formatting application for messages.

Extracted a SlackFormatter object, to contain formatting logic and
also external (user/channel) references, and pushed that object into
the Message object and its subsidiary objects, to allow the various
formattable values of the Attachment object to be formatted, or not
formatted, as appropriate.

* Bug fixes and extensions in reference insertion.

Many small changes to the special-markup code:

* removed slack-to-markdown translation and replaced with slack-to-html
  (so that fields that contain slack escapes but not markdown can be
  handled cleanly)
* made channel mention code handle both channels that do not start with
  "0" and annotated mentions
* made hyperlink code handle spaces in link text
* made hashtag code apply only to hashtags that start with a letter,
  to avoid picking up standard usage number printing (e.g. "#1")

* Removed custom em/strong handling.

Replaced custom bold/italic handling with slightly less aggressive
slack-to-markdown translation: this solution is imperfect but seems
to have fewer bugs than the manual processing approach.

* moved emoji handling before markdown processing to avoid markdown
  breaking emoji (since unicode emoji are valid markdown input)
* removed custom * and _ handlers, replacing them with a simple
  translator from "*" to "**", to get markdown bold from slack bold
* turned off "code-friendly" extra in markdown2 processing, so that
  bold and italics are once again provided by that step.

* Consolidated mention code.

Made all mentions, annotated or otherwise, hit the same code path.

* Precompile regular expressions.

* Refactored message/formatter creation.

Extracted SlackFormatter into its own file, and created it outside of
the Message constructor, to avoid the (incorrect) appearance of creating
a million independent SlackFormatter objects when actually they all just
wrap the same dicts.

* Standardized docstring format.

Made non-docstring-like strings into docstring-like strings, and
made one of them somewhat more informative.